backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
Pillow==11.1.0
httpx==0.28.1
types-python-jose
types-passlib